indexer/requirements.txt
opensearch-py==2.2.0
pandas==1.5.3
python-dotenv==1.0.0
retry==0.9.2